+++ src/main/java/cc/bitky/jetbrains/plugin/universalgenerate/action/actiongroup/GenerateAnnotationActionGroup.java
@@ -24,7 +24,7 @@ public class GenerateAnnotationActionGroup extends AbstractBitkylinUniversalGene
 
     @Override
     public AnAction @NotNull [] getChildren(@Nullable AnActionEvent anActionEvent) {
-        if (!GlobalSettingsStateHelper.getInstance().isContextMenuShowed()) {
+        if (!GlobalSettingsStateHelper.getInstance().isRightClickMenuEnabled()) {
             anActionEvent.getPresentation().setVisible(false);
             return new AnAction[0];
         }