+++ src/main/java/cc/bitky/jetbrains/plugin/universalgenerate/action/actiongroup/SwaggerToJavaDocActionGroup.java
@@ -23,7 +23,7 @@ public class SwaggerToJavaDocActionGroup extends AbstractBitkylinUniversalGenera
 
     @Override
     public AnAction @NotNull [] getChildren(@Nullable AnActionEvent anActionEvent) {
-        if (!GlobalSettingsStateHelper.getInstance().isContextMenuShowed()) {
+        if (!GlobalSettingsStateHelper.getInstance().isRightClickMenuEnabled()) {
             anActionEvent.getPresentation().setVisible(false);
             return new AnAction[0];
         }